requirements.txt
Flask
biopython
